Make Room!
I moved the "Potential" label over to the left, on top of "Field".  this is actually somewhat better because on the left boundary the voltage curve looks like it's to the left of the field curve so it makes it just a little more intuitive which is which.  more importantly it gives the timer room to breathe.  I left-alined it since this font apparently doesn't have monospace digits.