a flash of light
as I introduce more transient effects, I realized it's annoyingly hard to know when the video is looping.  I notice this with other similar graphics as well.  my solution is to add a 0.1s flash of white at the very end.  it may be a bit obnoxius, but if it only happens once every 8 seconds it shouldn't be too bad.  nice and intuitive!